src/requirements.txt
# AWS SDK
boto3==1.34.23
botocore==1.34.23

# JWT認証
PyJWT==2.8.0
cryptography==42.0.1

# バリデーション
pydantic==2.5.3

# 日時処理
python-dateutil==2.8.2

# OpenAI API
openai==1.7.2

# LINE Messaging API
line-bot-sdk==3.7.0

# HTTPリクエスト
requests==2.31.0

# 環境変数管理
python-dotenv==1.0.0

# ローカル開発用
Flask==3.0.0
flask-cors==4.0.0

# テスト
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
moto==4.2.11  # AWS モック

# 開発用
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2